requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
jinja2>=3.1
python-multipart>=0.0.9
keras>=3.0.0
tensorflow>=2.15.0
Flask>=3.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.26.0
h5py>=3.10.0